% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_em_model.R
\name{estimate_em_model}
\alias{estimate_em_model}
\title{Fit a Gamma Frailty Model for Interval Censored Data}
\usage{
estimate_em_model(
  d,
  s,
  Ci,
  Xp,
  n_knots = 5,
  spline_order = 2,
  tol = 1e-05,
  max_iter = 1500
)
}
\arguments{
\item{d}{A numeric vector for the event status indicator (1 if event present, 0 otherwise).}

\item{s}{A numeric vector for the informative censoring indicator (1 if informative censoring, 0 otherwise).}

\item{Ci}{A numeric vector of the observed censoring times.}

\item{Xp}{A numeric matrix of covariates. Each column is a covariate and each row is an observation.}

\item{n_knots}{An integer specifying the number of interior knots for the splines. Default is 5.}

\item{spline_order}{An integer specifying the order of the I-splines and M-splines. Default is 2.}

\item{tol}{The convergence tolerance for the EM algorithm. The algorithm stops when the maximum absolute difference between parameter estimates in consecutive iterations is less than this value. Default is 1e-5.}

\item{max_iter}{The maximum number of iterations for the EM algorithm. Default is 1500}
}
\description{
Estimates parameters for a Gamma frailty proportional hazards model using an
Expectation-Maximization (EM) algorithm. The model is designed for current
status data subject to a mix of informative and non-informative censoring.
}
