% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ispline.R
\name{Ispline}
\alias{Ispline}
\title{Generate I-Spline Basis Functions}
\usage{
Ispline(x, order, knots)
}
\arguments{
\item{x}{A numeric vector of values at which to evaluate the spline basis.}

\item{order}{An integer specifying the order of the spline (e.g., 2 for quadratic).}

\item{knots}{A numeric vector of knot locations for the spline.}
}
\value{
A matrix where each column is an I-spline basis function evaluated at the points in `x`.
}
\description{
Computes the basis functions for I-splines, which are integrated M-splines.
I-splines are non-decreasing functions useful for modeling cumulative hazard functions.
This function is based on the work of Ramsay (1988).
}
