% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Louis.R
\name{Louis}
\alias{Louis}
\title{Calculate Variance-Covariance Matrix using Louis's Method}
\usage{
Louis(t, Cb2, Cg2, Tb2, Tg2, bdervativeCi, bCi, s, d, Xp)
}
\arguments{
\item{t}{The final estimate for the frailty parameter `tau`.}

\item{Cb2}{Final estimates for the regression coefficients `beta_C`.}

\item{Cg2}{Final estimates for the spline coefficients `gamma_C`.}

\item{Tb2}{Final estimates for the regression coefficients `beta_T`.}

\item{Tg2}{Final estimates for the spline coefficients `gamma_T`.}

\item{bdervativeCi}{The M-spline basis matrix for the censoring times.}

\item{bCi}{The I-spline basis matrix for the censoring times.}

\item{s}{The event indicator for informative censoring (1 if event, 0 if censored).}

\item{d}{The tumor status indicator (1 if tumor present, 0 otherwise).}

\item{Xp}{The matrix of covariates.}
}
\value{
The estimated variance-covariance matrix for all model parameters.
}
\description{
Computes the observed information matrix to estimate the variance-covariance
matrix of the model parameters using the method described by Louis (1982).
This method is applied after the EM algorithm has converged.
}
