% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mspline.R
\name{Mspline}
\alias{Mspline}
\title{Generate M-Spline Basis Functions}
\usage{
Mspline(x, order, knots)
}
\arguments{
\item{x}{A numeric vector of values at which to evaluate the spline basis.}

\item{order}{An integer specifying the order of the spline (e.g., 2 for quadratic).}

\item{knots}{A numeric vector of knot locations for the spline.}
}
\value{
A matrix where each column is an M-spline basis function evaluated at the points in `x`.
}
\description{
Computes the basis functions for M-splines, which are non-negative splines
useful for modeling hazard functions. [cite_start]This function is based on the work of Ramsay (1988)[cite: 95].
}
